fix(build-audit): move schema into separate dir
